# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Creative Studio CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      parameters:
        - name: photographerId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a new lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Created lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
...
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeadBody"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /leads/{id}/convert:
    post:
      operationId: convertLead
      tags: [leads]
      summary: Convert a lead to a client + project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversion result with new client and project IDs
          content:
            application/json:
              schema:
                type: object
                properties:
                  clientId:
                    type: integer
                  projectId:
                    type: integer
                  message:
                    type: string
                required:
                  - clientId
                  - projectId
                  - message
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        estimatedValue:
          type: ["number", "null"]
        source:
          type: string
        status:
          type: string
          enum: [new, contacted, proposal_sent, negotiating, won, lost]
        notes:
          type: ["string", "null"]
        projectName:
          type: ["string", "null"]
        serviceId:
          type: ["integer", "null"]
        serviceName:
          type: ["string", "null"]
        lostReason:
          type: ["string", "null"]
        wonMonth:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - source
        - status
        - createdAt
    CreateLeadBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        estimatedValue:
          type: ["number", "null"]
        source:
          type: string
        status:
          type: string
          enum: [new, contacted, proposal_sent, negotiating, won, lost]
        notes:
          type: ["string", "null"]
        projectName:
          type: ["string", "null"]
        serviceId:
          type: ["integer", "null"]
        lostReason:
          type: ["string", "null"]
        wonMonth:
          type: ["string", "null"]
      required:
        - name
        - source
    UpdateLeadBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        estimatedValue:
          type: ["number", "null"]
        source:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        projectName:
          type: ["string", "null"]
        serviceId:
          type: ["integer", "null"]
        lostReason:
          type: ["string", "null"]
        wonMonth:
          type: ["string", "null"]
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        originalIdea:
          type: ["string", "null"]
        aiGeneratedIdea:
          type: ["string", "null"]
        proposedIdea:
          type: ["string", "null"]
        photographerId:
          type: ["integer", "null"]
        photographerName:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
        loginUsername:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - createdAt
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        originalIdea:
          type: ["string", "null"]
        aiGeneratedIdea:
          type: ["string", "null"]
        proposedIdea:
          type: ["string", "null"]
        photographerId:
          type: ["integer", "null"]
        password:
          type: ["string", "null"]
      required:
        - name
    UpdateClientBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        originalIdea:
          type: ["string", "null"]
        aiGeneratedIdea:
          type: ["string", "null"]
        proposedIdea:
          type: ["string", "null"]
        photographerId:
          type: ["integer", "null"]
...